+++ openstack/blockstorage/extensions/storagepools/testing/fixtures.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/gophercloud/gophercloud/openstack/blockstorage/storagepools"
+	"github.com/gophercloud/gophercloud/openstack/blockstorage/extensions/storagepools"
 	"github.com/gophercloud/gophercloud/testhelper"
 	"github.com/gophercloud/gophercloud/testhelper/client"
 )
