+++ openstack/blockstorage/extensions/storagepools/testing/requests_test.go
@@ -3,7 +3,7 @@ package testing
 import (
 	"testing"
 
-	"github.com/gophercloud/gophercloud/openstack/blockstorage/storagepools"
+	"github.com/gophercloud/gophercloud/openstack/blockstorage/extensions/storagepools"
 	"github.com/gophercloud/gophercloud/testhelper"
 	"github.com/gophercloud/gophercloud/testhelper/client"
 )
